php/edittemplateproperties.php: Added code to support editing of template properties.  The script now loads in the details of a template if editing the template.  Note that it's not yet possible to edit templates as it is not yet called from the template editor.
Added a function to convert a time given in 
seconds into an hh:mm formatted string, needed for turning the repeat 
times back into hours and minutes.

also added the ability to initialise a new template with settings from 
an existing template, with the option of also including the events.  The 
latter has not yet been implemented as it needs to be done in 
updatetemplate.php.


git-svn-id: svn://svn.code.sf.net/p/prsradio/code/trunk@398 410355f2-c4ea-45de-a5b5-301721f902b9