src/shoutmixeroutput.c shout_thread(): Another cast to prevent type warnings.
The same query regarding the non-void warning  applying to an apparently 
void function also applies here.

Note: The reconnect delay logic needs to be looked at here.  Right now, 
even though it can be specified in the config file, a value of 10 
seconds appears to be hard-coded.  Also this delay is only triggered 
when an attempt to connect fails.  It is not triggered when something 
else causes us to disconnect, such as an encoder dying.  since this has 
been happening during testing of this code, I can tell you this causes 
rapid connects/disconnects.  This would also seem to suggest that the 
death of an encoder needs looking at also.


git-svn-id: svn://svn.code.sf.net/p/prsradio/code/trunk@436 410355f2-c4ea-45de-a5b5-301721f902b9