Force madplay to give us a stereo stream for MP3 files encoded as dual channel.  This fixes the bug where some files would play at double speed.  Madplay's default when playing dual-channel files is to play the left channel only.  Since PRS is expecting stereo PCM output from madplay in all cases except a mono MP3 file, it plays the mono PCM data it receives at double its intended speed.
We now send an extra parameter to madplay to ensure we get the PCM 
format we expect.  We send "--stereo" in all cases except mono files, in 
which case we send "--mono".

In order to be able to send this extra parameter to madplay, I have had 
to change the mp3_decoder_new() function so as to take an extra argument 
- the number of channels expected.  This results in changes to 
mp3_decoder_new() in mp3decoder.c and mp3decoder.h, and to the functions 
which call it in mp3fileinfo.c and mp3mixerchannel.c.


git-svn-id: svn://svn.code.sf.net/p/prsradio/code/trunk@456 410355f2-c4ea-45de-a5b5-301721f902b9