Finally sorted out the licensing.  Licenses have been added at various stages, but they seem to have been lost along the way.
Added a COPYING file containing a license text as preferred by Marc 
Mulcahy, the main author of the PRS code.  This text was as a result of 
an Email exchange in September 2007.

Also updated the README file, clarifying the license and giving the 
license details for the resampler code which is licensed under the LGPL.


git-svn-id: svn://svn.code.sf.net/p/prsradio/code/trunk@460 410355f2-c4ea-45de-a5b5-301721f902b9